backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlalchemy-utils==0.42.1
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.0
Pillow==10.1.0
numpy==1.26.2
rasterio==1.3.9
shapely==2.0.2
geojson==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aiofiles==23.2.1
cryptography==41.0.7
pytz==2023.3
dateparser==1.1.8
